refactor(cli): split workflow debug handlers
