+++ packages/rust/biors/src/cli/mod.rs
@@ -1,11 +1,19 @@
 pub mod args;
 mod batch;
+mod debug;
+mod diff;
 mod doctor;
 pub mod handlers;
+mod pipeline;
+mod workflow;
 pub use args::{
     BatchCommand, Cli, Command, FastaCommand, KindArg, PackageCommand, PaddingArg, SeqCommand,
     TokenizerCommand, TokenizerProfileArg,
 };
 pub(crate) use batch::run_batch_command;
+pub(crate) use debug::run_debug;
+pub(crate) use diff::run_diff;
 pub(crate) use doctor::build_doctor_report;
 pub use handlers::run;
+pub(crate) use pipeline::run_pipeline;
+pub(crate) use workflow::run_workflow;
